refact: 캡슐 facade 적용
